+++ src/module/applications/api/application.mjs
@@ -1,7 +1,8 @@
 const { HandlebarsApplicationMixin, Application } = foundry.applications.api;
 
 /**
- * A stock application meant for async behavior using templates.
+ * A stock form application meant for async behavior using templates.
+ * @abstract
  */
 export default class DSApplication extends HandlebarsApplicationMixin(Application) {
   /** @inheritdoc */
@@ -15,6 +16,7 @@ export default class DSApplication extends HandlebarsApplicationMixin(Applicatio
       width: 450,
       height: "auto",
     },
+    timeout: null,
     tag: "form",
     window: {
       contentClasses: ["standard-form"],
@@ -53,6 +55,25 @@ export default class DSApplication extends HandlebarsApplicationMixin(Applicatio
     application.addEventListener("close", () => resolve(application.config), { once: true });
     application.render({ force: true });
     return promise;
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  async _onFirstRender(context, options) {
+    const timeout = this.options.timeout;
+
+    if (timeout) setTimeout(() => {
+      ui.notifications.error("DRAW_STEEL.SOCKET.WARNING.Timeout", {
+        format: {
+          label: this.constructor.name,
+          seconds: timeout / 1000,
+        },
+      });
+      this.close();
+    }, timeout);
+
+    await super._onFirstRender(context, options);
   }
 
   /* -------------------------------------------------- */