+++ src/module/applications/sheets/npc.mjs
@@ -1,5 +1,6 @@
 import { systemID, systemPath } from "../../constants.mjs";
 import DrawSteelActorSheet from "./actor-sheet.mjs";
+import { DocumentSourceInput, MonsterMetadataInput } from "../apps/_module.mjs";
 
 /** @import { FormSelectOption } from "@client/applications/forms/fields.mjs" */
 /** @import DrawSteelActor from "../../documents/actor.mjs"; */
@@ -174,7 +175,7 @@ export default class DrawSteelNPCSheet extends DrawSteelActorSheet {
    * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
    */
   static async #updateSource(event, target) {
-    this.actor.system.source.updateDialog();
+    new DocumentSourceInput({ document: this.document }).render({ force: true });
   }
 
   /**
@@ -184,45 +185,7 @@ export default class DrawSteelNPCSheet extends DrawSteelActorSheet {
    * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
    */
   static async #editMonsterMetadata(event, target) {
-    const htmlContainer = document.createElement("div");
-    const schema = this.actor.system.schema;
-    const monsterData = this.actor.system.monster;
-    const monsterConfig = ds.CONFIG.monsters;
-
-    const keywordInput = schema.getField("monster.keywords").toFormGroup({}, {
-      value: monsterData.keywords,
-      options: Object.entries(monsterConfig.keywords).map(([value, { label, group }]) => ({ value, label, group })),
-    });
-    const levelInput = schema.getField("monster.level").toFormGroup({}, { value: monsterData.level });
-    const organizationInput = schema.getField("monster.organization").toFormGroup({}, {
-      value: monsterData.organization,
-      options: Object.entries(monsterConfig.organizations).map(([value, { label }]) => ({ value, label })),
-    });
-    const roleInput = schema.getField("monster.role").toFormGroup({}, {
-      value: monsterData.role,
-      options: Object.entries(monsterConfig.roles).map(([value, { label }]) => ({ value, label })),
-    });
-    const evInput = schema.getField("monster.ev").toFormGroup({}, { value: monsterData.ev });
-
-    htmlContainer.append(keywordInput, levelInput, organizationInput, roleInput, evInput);
-
-    /** @type {object | null} */
-    const fd = await ds.applications.api.DSDialog.input({
-      content: htmlContainer.outerHTML,
-      classes: ["draw-steel", "monster-metadata"],
-      window: {
-        title: "DRAW_STEEL.Actor.NPC.MonsterMetadata.DialogTitle",
-        icon: "fa-solid fa-spaghetti-monster-flying",
-      },
-      ok: {
-        label: "Save",
-        icon: "fa-solid fa-floppy-disk",
-      },
-      rejectClose: false,
-    });
-    if (fd) {
-      await this.actor.update(fd);
-    }
+    new MonsterMetadataInput({ document: this.document }).render({ force: true });
   }
 
   /**
@@ -232,6 +195,11 @@ export default class DrawSteelNPCSheet extends DrawSteelActorSheet {
    * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
    */
   static async #freeStrike(event, target) {
+    try { game.user.targets.map(t => t.actor); } catch (e) {
+      ui.notifications.error("DRAW_STEEL.Actor.NPC.FreeStrike.MultiLinked", { localize: true });
+      throw (e);
+    }
+
     /** @type {Array<DrawSteelActor>} */
     const targets = game.user.targets.map(t => t.actor).filter(a => a?.system?.takeDamage).toObject();
     if (!targets.length) {
@@ -254,6 +222,7 @@ export default class DrawSteelNPCSheet extends DrawSteelActorSheet {
       const formGroup = foundry.applications.fields.createFormGroup({
         label: a.name,
         input: checkboxInput,
+        classes: ["inline"],
       });
       // style fix
       const label = formGroup.querySelector("label");