+++ src/module/data/item/base.mjs
@@ -8,7 +8,6 @@ const fields = foundry.data.fields;
 
 /**
  * A base item model that provides basic description and source metadata for an item instance
- * @extends foundry.abstract.TypeDataModel
  */
 export default class BaseItemModel extends SubtypeModelMixin(foundry.abstract.TypeDataModel) {
   /**
@@ -36,8 +35,9 @@ export default class BaseItemModel extends SubtypeModelMixin(foundry.abstract.Ty
 
     /**
      * The Draw Steel ID, indicating a unique game rules element
+     * @remarks `readonly: true` makes this non-iterable
      */
-    schema._dsid = new fields.StringField({ blank: false });
+    schema._dsid = new fields.StringField({ required: true, readonly: true });
 
     return schema;
   }