+++ src/module/data/message/_module.mjs
@@ -1,2 +1,3 @@
 export { default as BaseMessageModel } from "./base.mjs";
 export { default as AbilityUseModel } from "./ability-use.mjs";
+export { default as SavingThrowModel } from "./saving-throw.mjs";