+++ src/module/data/pseudo-documents/_types.d.ts
@@ -1,4 +1,4 @@
-import "./power-roll-effects/_types"
+import "./power-roll-effects/_types";
 
 import { ApplicationConfiguration } from "@client/applications/_types";
 import { DialogV2Configuration, DialogV2WaitOptions } from "@client/applications/api/dialog.mjs";