+++ src/module/data/settings/hero-tokens.mjs
@@ -35,10 +35,10 @@ export class HeroTokenModel extends foundry.abstract.DataModel {
   /**
    * Send a socket message to the Director to spend a hero token
    * Necessary because only game masters can modify world settings
-   * @param {string} spendType       Key of `ds.CONFIG.hero.tokenSpends`.
-   * @param {object} [options]       Options to modify the token spend
+   * @param {string} spendType        Key of `ds.CONFIG.hero.tokenSpends`.
+   * @param {object} [options]        Options to modify the token spend
    * @param {string} [options.flavor] Flavor for the chat message (default: Current user's character name)
-   * @returns {Promise<void|false>}           An explicit `false` if there was an error in spending the token
+   * @returns {Promise<void|false>}   An explicit `false` if there was an error in spending the token
    */
   async spendToken(spendType, options = {}) {
     if (!game.users.activeGM) {