+++ src/module/documents/active-effect.mjs
@@ -53,13 +53,11 @@ export default class DrawSteelActiveEffect extends foundry.documents.ActiveEffec
   /**
    * Automatically deactivate effects with expired durations
    * @inheritdoc
-   * @type {Boolean}
    */
   get isSuppressed() {
-    if (Number.isNumeric(this.duration.remaining)) {
-      return this.duration.remaining <= 0;
-    }
-    return false;
+    if (Number.isNumeric(this.duration.remaining)) return this.duration.remaining <= 0;
+    // Checks `system.isSuppressed`
+    else return super.isSuppressed;
   }
 
   /** @inheritdoc */