+++ src/module/documents/base-document-mixin.mjs
@@ -1,6 +1,10 @@
+/** @import Document from "@common/abstract/document.mjs"; */
+
 /**
  * Mixin for common functions used across most or all document classes in this system.
- * @mixin
+ * Requires the document to have a `system` field.
+ * @template {import("@common/_types.mjs").Constructor<Document>} BaseDocument
+ * @param {BaseDocument} base
  *
  */
 export default base => {