+++ CHANGELOG.md
@@ -18,6 +18,33 @@
 ### Known Issues
 -->
 
+### 0.7.2
+
+### Added
+- Added support for the automatic end of effects. (#551)
+  - End of Turn will automatically self-disable when an affected actor ends their turn.
+  - Saving Throws will create prompts for owners. If multiple players own an actor, the active GM will receive a dialog to help delegate rolls.
+    - The roll message has a button to spend a hero token to automatically succeed.
+  - End of Encounter effects will automatically self-disable alongside the encounter.
+- Added Disengage to Stats tab in the movement section and integrated the Kit Bonus. (#519)
+- Added UI for configuring an actor's turns per round, save bonus, and save threshold. (#532)
+  - These are also readable in a hover tooltip on the "Combat" legend in the Stats tab.
+- Added spend line to the displayed power content. (#398)
+  - Note: Expect future breaking changes with how Spend is configured to account for variations that the current data does not represent.
+- Added Ability compendium with basic actions and maneuvers.
+  - Creating a new character will give them those basic actions and maneuvers. (#56)
+- Stamina changes now display as floating damage number. The color depends on the type of damage dealt. (#417)
+
+### Changed
+- Aligned the system's active effect suppression checks with core v13 behavior, allowing the system model to also apply. (#541)
+- The Monster Metadata and Source forms now perform live updates instead of using a save button.
+- Added a `dsid` getter on the Item class, made the initialized version of _dsid read only (you can still perform updates but not apply active effects). (#389)
+
+### Fixed
+- Fixed an issue that allowed characteristics to become null, instead of defaulting back to 0.
+- Fixed free strikes not applying. (#544)
+- Minions which have unlinked tokens in a squad but are not themselves in a squad will not display as if they are in a squad. (#561)
+
 ## 0.7.1
 
 ### Added